boarddocs_loader/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boarddocs-loader"
version = "0.1.0"
description = "Data loader for Kent School District BoardDocs meeting records"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "tqdm>=4.66.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
